Fix Product.BoxWeight type

+++ Core/Entities/Product.cs
@@ -6,7 +6,7 @@ namespace Core.Entities
 		public int BoxLenght { get; set; }
 		public int BoxHeight { get; set; }
 		public int BoxWidth { get; set; }
-		public int BoxWeight { get; set; }
+		public double BoxWeight { get; set; }
 		public int Dose { get; set; }
 		public string DoseControl { get; set; }
 		public double Speed { get; set; }
